requirements.txt
# Production Requirements for Azure Deployment
# Optimized for fast deployment - ML dependencies are OPTIONAL
# The VPN optimizer auto-detects ML availability and adapts accordingly

# FastAPI and ASGI server
fastapi==0.115.12
uvicorn[standard]==0.32.1
gunicorn==21.2.0
python-multipart==0.0.9

# Database
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
cryptography==42.0.8
pyotp==2.9.0

# Payment Processing
stripe==10.16.0
httpx==0.27.0

# VPN & Networking
qrcode==7.4.2
pillow==10.4.0

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.0
email-validator==2.2.0

# Security & Rate Limiting
slowapi==0.1.9
redis==5.0.1

# Azure Services (optional - for production)
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Monitoring & Logging (Section 9)
sentry-sdk==1.40.0
opencensus-ext-azure==1.1.9
applicationinsights==0.11.10
psutil==5.9.8

# Email Services (Section 7)
sendgrid==6.11.0
boto3==1.34.0
jinja2==3.1.2

# DNS Operations (Section 8)
dnspython==2.4.2

# NOTE: ML packages (xgboost, numpy, scikit-learn) are OPTIONAL
# To enable ML-enhanced optimization, install requirements_ml.txt separately
# The app works perfectly without them using MARL-only mode
